Cleanup SQL series
